Prepared stuff for the ai.h
