+++ poCker-tests.depend
@@ -8,7 +8,7 @@
 	<math.h>
 	"CuTest.h"
 
-1714903044 c:\users\ja\documents\github\pocker\test\cutest.h
+1715178291 c:\users\ja\documents\github\pocker\test\cutest.h
 	<setjmp.h>
 	<stdarg.h>
 
@@ -23,10 +23,10 @@
 	"CuTest.h"
 	"../src/utils.h"
 
-1714934589 c:\users\ja\documents\github\pocker\src\utils.h
+1715177686 c:\users\ja\documents\github\pocker\src\utils.h
 	<stdbool.h>
 
-1714934589 source:c:\users\ja\documents\github\pocker\src\utils.c
+1715177645 source:c:\users\ja\documents\github\pocker\src\utils.c
 	"utils.h"
 	<ctype.h>
 	<stdarg.h>
@@ -35,7 +35,7 @@
 	<stdlib.h>
 	<string.h>
 
-1714903044 source:c:\users\ja\documents\github\pocker\src\handranking.c
+1715166194 source:c:\users\ja\documents\github\pocker\src\handranking.c
 	"handranking.h"
 	<limits.h>
 	<math.h>
@@ -44,7 +44,7 @@
 	"playingcard.h"
 	"utils.h"
 
-1714934589 c:\users\ja\documents\github\pocker\src\playingcard.h
+1715177902 c:\users\ja\documents\github\pocker\src\playingcard.h
 	<string.h>
 
 1713438951 source:c:\users\ja\documents\github\pocker\test\test_handranking.c
@@ -53,7 +53,7 @@
 	"CuTest.h"
 	"../src/playingcard.h"
 
-1714934589 source:c:\users\ja\documents\github\pocker\src\dealer.c
+1715163094 source:c:\users\ja\documents\github\pocker\src\dealer.c
 	"dealer.h"
 	<stdbool.h>
 	<stdio.h>
@@ -68,14 +68,14 @@
 1713035139 c:\users\ja\documents\github\pocker\src\handranking.h
 	"player.h"
 
-1715160760 c:\users\ja\documents\github\pocker\src\player.h
+1715175855 c:\users\ja\documents\github\pocker\src\player.h
 	<stdbool.h>
 	"playingcard.h"
 	"gamerules.h"
 
 1713035338 c:\users\ja\documents\github\pocker\src\playingcard_enums.h
 
-1714934589 source:c:\users\ja\documents\github\pocker\src\playingcard.c
+1715178099 source:c:\users\ja\documents\github\pocker\src\playingcard.c
 	"playingcard.h"
 	<stdio.h>
 	<string.h>
@@ -83,7 +83,7 @@
 1712917163 source:c:\users\ja\documents\github\pocker\src\constants.c
 	"constants.h"
 
-1715108361 source:c:\users\ja\documents\github\pocker\src\player.c
+1715175878 source:c:\users\ja\documents\github\pocker\src\player.c
 	"player.h"
 	<stdlib.h>
 
@@ -93,28 +93,29 @@
 	"../src/dealer.h"
 	"../src/player.h"
 
-1714903044 c:\users\ja\documents\github\pocker\src\dealer.h
+1715162599 c:\users\ja\documents\github\pocker\src\dealer.h
 	<stdbool.h>
 	"gamerules.h"
 	"gamestate.h"
 	"player.h"
 	"playingcard.h"
 
-1715160598 c:\users\ja\documents\github\pocker\src\gamerules.h
+1715164031 c:\users\ja\documents\github\pocker\src\gamerules.h
 	<stdbool.h>
 
-1715160547 source:c:\users\ja\documents\github\pocker\src\gamerules.c
+1715163658 source:c:\users\ja\documents\github\pocker\src\gamerules.c
 	"gamerules.h"
 	<math.h>
 	<stdio.h>
 	"messages.h"
 
-1715160598 source:c:\users\ja\documents\github\pocker\src\gamestate.c
+1715178386 source:c:\users\ja\documents\github\pocker\src\gamestate.c
 	"gamestate.h"
 	<stdbool.h>
 	<stdlib.h>
 	<stdio.h>
 	<math.h>
+	"ai.h"
 	"dealer.h"
 	"gamerules.h"
 	"messages.h"
@@ -122,7 +123,7 @@
 	"player.h"
 	"utils.h"
 
-1715160598 c:\users\ja\documents\github\pocker\src\gamestate.h
+1715166006 c:\users\ja\documents\github\pocker\src\gamestate.h
 	<stdbool.h>
 	"gamerules.h"
 	"player.h"
@@ -316,15 +317,15 @@
 	<string.h>
 	<limits.h>
 
-1715104615 c:\users\ja\documents\github\pocker\src\messages.h
+1715177567 c:\users\ja\documents\github\pocker\src\messages.h
 
-1715099455 c:\users\ja\documents\github\pocker\src\playerio.h
+1715176036 c:\users\ja\documents\github\pocker\src\playerio.h
 	<stdbool.h>
 	"gamerules.h"
 	"gamestate.h"
 	"player.h"
 
-1715109658 source:c:\users\ja\documents\github\pocker\src\playerio.c
+1715178221 source:c:\users\ja\documents\github\pocker\src\playerio.c
 	"playerio.h"
 	<ctype.h>
 	<limits.h>
@@ -338,7 +339,7 @@
 	"gamestate.h"
 	"utils.h"
 
-1714934589 source:c:\users\ja\documents\github\pocker\src\messages.c
+1715177374 source:c:\users\ja\documents\github\pocker\src\messages.c
 	"messages.h"
 	<stdio.h>
 	<string.h>
@@ -352,9 +353,24 @@
 	<stdbool.h>
 	<stdio.h>
 
-1714903044 source:c:\users\ja\documents\github\pocker\test\test_playerio.c
+1715178327 source:c:\users\ja\documents\github\pocker\test\test_playerio.c
 	<limits.h>
 	"CuTest.h"
 	"../src/playerio.h"
 	"../src/messages.h"
+
+1715178293 source:c:\users\ja\documents\github\pocker\test\test_ai.c
+	"CuTest.h"
+	"../src/ai.h"
+
+1715162599 c:\users\ja\documents\github\pocker\src\ai.h
+	"player.h"
+	"gamestate.h"
+	"gamerules.h"
+
+1715175855 source:c:\users\ja\documents\github\pocker\src\ai.c
+	"ai.h"
+	"player.h"
+	"gamestate.h"
+	"gamerules.h"
