+++ src/player.c
@@ -25,13 +25,6 @@ Player* playerCreateNewWithFunds(int initial_funds){
 }
 
 /**
- *  \brief TODO
- */
- int takeAction(Player *self){
-    return 0;
- }
-
-/**
  *  \brief Resets player's scores to an array of zeros.
  *
  *  Remember that C does not have zero-initialization. The scores array WILL be filled with gibberish data.
